bot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
python-telegram-bot==13.15
pytz==2025.1
setuptools==75.8.0
six==1.17.0
tornado==6.1
tzlocal==5.2
urllib3==2.3.0
python-dotenv
requests
